OWASPVerificator/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
jinja2==3.1.6
sqlalchemy==2.0.40
pymysql==1.1.1
python-dotenv==1.1.0
requests==2.32.3
python-multipart==0.0.20
pytest==8.3.5
httpx==0.28.1
